:recycle: Refactored color-tokens setup
