--- @navikt/core/tokens/darkside/tokens/colors/index.ts
@@ -1,4 +1,0 @@
-export {
-  globalColorDarkModeConfig,
-  globalColorLightModeConfig,
-} from "./global-config";
