+++ @navikt/core/tokens/darkside/tokens/colors/semantic-roles.tokens.ts
@@ -1,12 +1,11 @@
-import _ from "lodash";
-import { ColorRolesList, type SemanticColorRoles } from "../../../types";
+import { type SemanticColorRoles } from "../../../types";
 import { type StyleDictionaryTokenConfig } from "../../tokens.util";
 
 /**
- * Gray colors are percieved a little lighter than other colored versions,
- * so we make some adjustments for neutral colors.
+ * Maps the semantic colors to the global color layer for a given role.
+ * @note Gray is handled a little differently, as it is visually perceived a little lighter than other colors.
  */
-export const configForRole = (role: SemanticColorRoles) => {
+export function semanticRoleTokens(role: SemanticColorRoles) {
   return {
     bg: {
       [`${role}-soft`]: {
@@ -117,97 +116,4 @@ export const configForRole = (role: SemanticColorRoles) => {
       },
     },
   } satisfies StyleDictionaryTokenConfig<"color">;
-};
-
-export const themedConfigForRole = (role: SemanticColorRoles) => {
-  return {
-    bg: {
-      soft: {
-        value: `{ax.bg.${role}-soft.value}`,
-        type: "themed-role",
-      },
-      softA: {
-        value: `{ax.bg.${role}-softA.value}`,
-        type: "themed-role",
-      },
-      moderate: {
-        value: `{ax.bg.${role}-moderate.value}`,
-        type: "themed-role",
-      },
-      moderateA: {
-        value: `{ax.bg.${role}-moderateA.value}`,
-        type: "themed-role",
-      },
-      "moderate-hover": {
-        value: `{ax.bg.${role}-moderate-hover.value}`,
-        type: "themed-role",
-      },
-      "moderate-hoverA": {
-        value: `{ax.bg.${role}-moderate-hoverA.value}`,
-        type: "themed-role",
-      },
-      "moderate-pressed": {
-        value: `{ax.bg.${role}-moderate-pressed.value}`,
-        type: "themed-role",
-      },
-      "moderate-pressedA": {
-        value: `{ax.bg.${role}-moderate-pressedA.value}`,
-        type: "themed-role",
-      },
-      strong: {
-        value: `{ax.bg.${role}-strong.value}`,
-        type: "themed-role",
-      },
-      "strong-hover": {
-        value: `{ax.bg.${role}-strong-hover.value}`,
-        type: "themed-role",
-      },
-      "strong-pressed": {
-        value: `{ax.bg.${role}-strong-pressed.value}`,
-        type: "themed-role",
-      },
-    },
-    text: {
-      default: {
-        value: `{ax.text.${role}.value}`,
-        type: "themed-role",
-      },
-      subtle: {
-        value: `{ax.text.${role}-subtle.value}`,
-        type: "themed-role",
-      },
-      icon: {
-        value: `{ax.text.${role}-icon.value}`,
-        type: "themed-role",
-      },
-    },
-    border: {
-      default: {
-        value: `{ax.border.${role}.value}`,
-        type: "themed-role",
-      },
-      subtle: {
-        value: `{ax.border.${role}-subtle.value}`,
-        type: "themed-role",
-      },
-      subtleA: {
-        value: `{ax.border.${role}-subtleA.value}`,
-        type: "themed-role",
-      },
-      strong: {
-        value: `{ax.border.${role}-strong.value}`,
-        type: "themed-role",
-      },
-    },
-  };
-};
-
-/**
- * We need to deep merge the token config for each role to get the complete token config for all roles.
- */
-export const semanticTokensForAllRolesConfig = () => {
-  return ColorRolesList.reduce(
-    (acc, role) => _.merge(acc, configForRole(role)),
-    {} as ReturnType<typeof configForRole>,
-  );
-};
+}
